Extend vmray_submit to work with sample-only results
If do_not_reanalyze is set to true and a sample is submitted to
the Analyzer that it already knows, the API only returns the sample,
not a submission. These code changes enable the vmray_submit module
to also work with this kind of answers.
This commit also introduces a more robust parsing logic for
user-provided settings and unit tests for the vmray_submit module.